Provide the tty-layer mutex when initializing the pps api.  This allows time_pps_fetch() to be used in blocking mode.
Also, don't init the pps api for system devices (consoles) that provide a
custom attach routine.  The device may actually be a keyboard or other non-
tty device.  If it wants to do pps processing (unlikely) it must handle
everything for itself.  (In reality, only a sun keyboard uses a custom
attach routine, and it doesn't make a good pps device.)


git-svn-id: https://svn.freebsd.org/base/head@286469 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f