Direct commit
When compiling on -current from r284356 to r285986, the host's
crunchgen can't be used because builds of rescue break when STRIP= was
used to prevent stripping of binaries.  Direct commit because -current
code is different. Building 10 on current is common enough to warrnat
a safety belt since this has been broken for two months. The offending
change was never in stable/10.


git-svn-id: https://svn.freebsd.org/base/stable/10@286673 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f