Add a routine to return the hardware instance/unit number from ti,hwmods, given the hardware name.
The ti,hwmods property is used (among other things) to associate an fdt node
with a specific instance of some hardware.  For example given a device node
that contains the property ti,hwmods = "timer3", if you call this passing
"timer" as the hwmod string to look for it would return 3.


git-svn-id: https://svn.freebsd.org/base/head@286678 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f