Reinstate unify_tcp_port_space and associated code that was lost during the last OFED update (r278886).
iWARP on FreeBSD is properly integrated with the network stack and the
iWARP drivers _never_ operate out of any private TCP port-space that is
invisible to the kernel.  Instead, an iWARP connection shows up as a TCP
socket (which is what it is) fully visible to the kernel and standard
tools like netstat, sockstat, etc.


git-svn-id: https://svn.freebsd.org/base/head@286703 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f